% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweep_dataframe.R
\name{tweep_dataframe}
\alias{tweep_dataframe}
\title{Create dataframe of tweeps, either followers or people that you follow}
\usage{
tweep_dataframe(user_id, followers = TRUE)
}
\arguments{
\item{user_id}{character. Twitter id to query.}

\item{followers}{TRUE/FALSE depending on whether you want your followers or people you follow.}
}
\description{
Create dataframe of tweeps, either followers or people that you follow
}
