% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweep_map.R
\name{tweep_map}
\alias{tweep_map}
\title{Create global map of tweeps from output of tweep_dataframe}
\usage{
tweep_map(tweep_df, jitter = 2, ggplotly = TRUE, ...)
}
\arguments{
\item{tweep_df}{dataframe from \code{tweep_dataframe}}

\item{jitter}{numeric. How much to jitter the dots so they don't stack up on eachother.}

\item{ggplotly}{TRUE/FALSE depending on whether you want interactive output from ggplotly or not.}
}
\description{
Create global map of tweeps from output of tweep_dataframe
}
