% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweep_plot.R
\name{tweep_plot}
\alias{tweep_plot}
\title{plot a world map of your followers / people you follow}
\usage{
tweep_plot(user, followers = TRUE, jitter = 2, ggplotly = TRUE, ...)
}
\arguments{
\item{user}{character. Twitter id to query.}

\item{followers}{TRUE/FALSE depending on whether you want your followers or people you follow.}

\item{jitter}{numeric. How much to jitter the dots so they don't stack up on eachother.}

\item{ggplotly}{TRUE/FALSE depending on whether you want interactive output from ggplotly or not.}
}
\description{
plot a world map of your followers / people you follow
}
\details{
You need to register your twitter and google apis first. See the readme.
}
